+++ src/main/java/com/github/halvra/opencell/actions/EnvironmentComboBoxAction.java
@@ -2,14 +2,16 @@ package com.github.halvra.opencell.actions;
 
 import com.github.halvra.opencell.settings.ProjectSettingsState;
 import com.github.halvra.opencell.settings.model.Environment;
+import com.github.halvra.opencell.utils.ScriptUtil;
 import com.intellij.ide.DataManager;
-import com.intellij.openapi.actionSystem.AnAction;
-import com.intellij.openapi.actionSystem.CommonDataKeys;
-import com.intellij.openapi.actionSystem.DefaultActionGroup;
+import com.intellij.openapi.actionSystem.*;
 import com.intellij.openapi.actionSystem.ex.ComboBoxAction;
 import com.intellij.openapi.project.DumbAware;
+import com.intellij.openapi.project.IndexNotReadyException;
 import com.intellij.openapi.project.Project;
+import com.intellij.psi.PsiFile;
 import com.intellij.ui.components.panels.NonOpaquePanel;
+import com.intellij.ui.scale.JBUIScale;
 import com.intellij.util.ui.JBUI;
 import com.intellij.util.ui.UIUtil;
 import org.jetbrains.annotations.NotNull;
@@ -20,8 +22,39 @@ import java.awt.*;
 
 public abstract class EnvironmentComboBoxAction extends ComboBoxAction implements DumbAware {
     @Override
+    public void update(@NotNull AnActionEvent e) {
+        Presentation presentation = e.getPresentation();
+        PsiFile psiFile = e.getData(CommonDataKeys.PSI_FILE);
+
+        try {
+            if (!ScriptUtil.isScript(psiFile)) {
+                presentation.setEnabledAndVisible(false);
+            } else {
+                Project project = e.getData(CommonDataKeys.PROJECT);
+                presentation.setEnabled(project != null && !project.isDisposed() && project.isOpen());
+            }
+        } catch (IndexNotReadyException ignored) {
+            presentation.setEnabled(false);
+        }
+    }
+
+    @Override
     protected boolean shouldShowDisabledActions() {
         return true;
+    }
+
+    @NotNull
+    @Override
+    public JComponent createCustomComponent(@NotNull final Presentation presentation, @NotNull String place) {
+        ComboBoxButton myButton = new ComboBoxButton(presentation) {
+            @Override
+            public Dimension getPreferredSize() {
+                Dimension d = super.getPreferredSize();
+                d.width = Math.max(d.width, JBUIScale.scale(75));
+                return d;
+            }
+        };
+        return createCustomComponent(myButton);
     }
 
     @NotNull